added new class Component
